fix : 방 생성 API 로직 에러 수정

+++ backend/src/main/java/signiel/heartsigniel/model/member/dto/SignResponse.java
@@ -30,7 +30,7 @@ public class SignResponse {
 
     private boolean isBlack;
 
-    private int rating;
+    private Long rating;
 
     private int meetingCount;
